requirements.txt
django
djangorestframework
psycopg2-binary
python-decouple
pillow
Jinja2
drf-yasg